% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_smoother.R
\name{bt_eps}
\alias{bt_eps}
\title{Perform one Bootstrap Step}
\usage{
bt_eps(bt_rep, smoother_pts, resid, study_day, width)
}
\value{

}
\description{
Helper function to bootstrap the epsilon (error of AR) and then reconstruct
smoother (currently the moving median) using AR model and bootstrapped
epsilon
}
