% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_smoother.R
\name{bt_smoother}
\alias{bt_smoother}
\title{Bootstrap the Smoother}
\usage{
bt_smoother(smoother_pts, resid, study_day, width, bt_tot_rep)
}
\arguments{
\item{smoother_pts}{A data frame containing the smoother. Preferably the
output of one of the smoother functions included in this package.}

\item{resid}{A vector containing the residuals of the smoother to the data
points.}

\item{study_day}{A vector containing the study days at which the data points
were collected}

\item{width}{The width of the window for the moving median in days.}

\item{bt_tot_rep}{The amount of bootstrap repetitions to be performed.}
}
\value{

}
\description{
Fit an autoregressive model on the residuals of the smoother, then bootstrap
the error of the autoregressive model. Reconstruct the smoother by adding the
bootstrapped error, the fitted autoregressive model, and the smoother to form
the bootstrapped smoother.
}
\details{
An autoregressive (AR) model is used for the residuals of the smoother:
\deqn{Y(t) = S(t) + \eta(t)}
\deqn{\eta(t) = \sum^{p}_{j = 1} \phi_j \eta(t - j) + \epsilon}
where \eqn{t} is the point in time, \eqn{Y(t)} the data point,
\eqn{S(t)} a smoother, \eqn{\eta(t)} the residual of the smoother, \eqn{p}
the order of the AR model, \eqn{\phi_j} the coefficients of the AR model, and
\eqn{\epsilon} the error of the AR model.

The bootstrap procedure is as follows:
\enumerate{
\item Compute the smoother \eqn{S(t)}.
\item Compute the residuals \eqn{\eta(t_i) = Y(t_i) - S(t_i)}.
\item Fit an AR(p) model to \eqn{\eta(t_i)} to obtain the coefficients
\eqn{\phi_1, \dots, \phi_p} and residuals \eqn{\epsilon(t_i) = \eta(t_i) -
    \sum^{p}_{j = 1} \phi_j \eta(t_{i} - t_{i-j})}.
\item Resample \eqn{\epsilon(t_i)*} from \eqn{\epsilon(t_{p+1}), \dots,
    \epsilon(t_n)} to obtain \deqn{\eta(t_i)* = \sum^{p}_{j =1} \phi_j \eta(t_{i-j})*
    + \epsilon(t_{i - j})*} \deqn{Y(t_i)* = S(t_i) + \eta(t_i)*.}
\item Compute \eqn{S(.)* = g(Y(t_1), \dots, Y(t_n))} where \eqn{g} is the
function with which the smoother is calculated.
\item Repeat steps 4 and 5 \code{bt_tot_rep} times.
}
}
\references{
Bühlmann, P. (1998). Sieve Bootstrap for Smoothing in
Nonstationary Time Series. \emph{The Annals of Statistics}, 26(1), 48-83.
}
