% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_band.R
\name{conf_band}
\alias{conf_band}
\title{Confidence Bounds of the Smoother}
\usage{
conf_band(bt_smoother, smoother_pts, bt_tot_rep, conf_band_lvl)
}
\arguments{
\item{bt_smoother}{A data frame containing the bootstrapped smoother. Use the output of \code{bt_smoother}.}

\item{smoother_pts}{A data frame containing the smoother. Preferably the
output of one of the smoother functions included in this package.}

\item{bt_tot_rep}{The number of times we perform the bootstrap.}

\item{conf_band_lvl}{The confidence level for the simultaneous confidence
bands. Defaults to 0.95.}
}
\value{
A data frame containing the upper confidence bound, the lower confidence bound,
and the study day corresponding to the bounds.
}
\description{
Calculate the confidence bounds of the smoother function using the bootstrap.
}
\details{
The procedure is as follows:
\enumerate{
\item We compute the quantiles \deqn{ qₓ(tᵢ), q₁₋ₓ(tᵢ) i = 1,\dots, N}

    where \deqn{qₓ(tᵢ) = inf {u; P*[S(tᵢ)*ᵦ - S(tᵢ) \le u] \ge x} } is a
    pointwise bootstrap quantile, \eqn{S(tᵢ)*ᵦ the bootstrapped smoother},
    and \eqn{N} the number of measurements.
\item We vary the pointwise error \eqn{2x} until \deqn{P*[qₓ(tᵢ) \le S(tᵢ)*ᵦ - S(tᵢ) \le q₁₋ₓ(tᵢ) ⩝ i = 1,\dots, N] ≈ 1-\alpha.}
    In other words, until the ratio of bootstrap curves that have all their points within
\eqn{[qₓ(tᵢ), q₁₋ₓ(tᵢ)]} is approximately \eqn{1-\alpha}.
\item We let
\deqn{ Iₙ(tᵢ) = [S(tᵢ) +  qₓ(tᵢ), S(tᵢ) + q₁₋ₓ(tᵢ)] ⩝ i = 1, \dots, N}
  confidence bounds. Then \eqn{{Iₙ(tᵢ); i = 1,\dots, N}} is a consistent simultaneous confidence band of level \eqn{1-\alpha}.

}
}
\references{
Bühlmann, P. (1998). Sieve Bootstrap for Smoothing in
Nonstationary Time Series. \emph{The Annals of Statistics}, 26(1), 48-83.
}
