% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edecob.R
\name{edecob}
\alias{edecob}
\title{edecob}
\usage{
edecob(
  data,
  date,
  subj_id = "subj1",
  bt_tot_rep,
  learn_dur,
  basel_dur,
  event_min_dur = as.difftime(12, units = "weeks"),
  thresh_diff = 0.1,
  smoother = "mov_med",
  width = as.difftime(12, units = "weeks"),
  alpha = 0.05,
  plot = TRUE
)
}
\arguments{
\item{data}{A vector containing the data.}

\item{date}{A vector containing the dates at which the data points were collected}

\item{subj_id}{A string containing an identifier for the data.}

\item{bt_tot_rep}{The amount of bootstrap repetitions to be performed.}

\item{learn_dur}{The amount of days where the data is to be ignored.}

\item{basel_dur}{The width of the window from which the median is taken to use as baseline.}

\item{event_min_dur}{The minimal duration the worsening needs to be sustained to define an event.}

\item{thresh_diff}{The percentage difference between the baseline and threshold. A negative number indicates}

\item{smoother}{Which smoother is to be used. Currently only "mov_med" for moving median available.}

\item{width}{The width of the window for the moving median.}

\item{alpha}{The confidence level for the simultaneous confidence bands.}

\item{plot}{Whether a plot should be generated.}
}
\value{
A list of four values: \code{event_detected} gives whether an
event was detected, \code{event_detection_date} gives the date at which the
event was detected, \code{event_duration} gives the duration the event is
sustaind and \code{event_censored} gives whether the detected event is censored.
}
\description{
edecob
}
