% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edecob_plot.R
\name{edecob_plot}
\alias{edecob_plot}
\title{Plot Event Data}
\usage{
edecob_plot(
  data,
  study_day,
  smoother_pts,
  conf_band,
  event,
  subj_id = "subj1",
  learn_dur,
  basel_dur,
  thresh_diff = -0.1,
  width
)
}
\arguments{
\item{data}{A vector containing the data.}

\item{study_day}{A vector containing the study days at which the data points
were collected}

\item{smoother_pts}{A data frame containing the smoother. Preferably the
output of one of the smoother functions included in this package.}

\item{conf_band}{A data frame containing the confidence bounds. Ideally the output of \code{\link{conf_band}}.}

\item{event}{A list containing the events. Preferably the output of \code{\link{detect_event()}}}

\item{learn_dur}{The amount of days where the data is to be ignored.}

\item{basel_dur}{The width of the window in days from which the median is
taken to use as baseline.}

\item{thresh_diff}{The percentage difference between the baseline and
threshold. A negative number indicates a threshold below the baseline and
vice versa.}
}
\value{
A \code{ggplot2} object that visualizes the data.
}
\description{
Generates a ggplot2 visualizing the data and event if any detected.
}
\details{
The data points are plotted on the x-axis with the study day on the y-axis.
The data from the learning period are gray. The baseline period is outlined
by two vertical blue lines. The smoother is plotted in orange. The confidence
bound  is the blue area. If an event is detected, a red triangle will mark
the detection day.
}
\examples{

}
