% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Artificially generated data for examples}
\format{
A dataframe with 2100 rows and 5 variables:
\describe{
\item{subject}{The subject identifier}
\item{study_day}{The study day from which the data is collected}
\item{jump_height}{The jump height, in cm}
\item{baseline}{The baseline for the subject}
\item{threshold}{The threshold for the subject}
}
}
\usage{
example_data
}
\description{
A dataset in the long format containing the jump height
of 5 subjects over 350 days.
}
\keyword{datasets}
