% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_smoother.R
\name{fun_running_median}
\alias{fun_running_median}
\title{Moving Median}
\usage{
fun_running_median(win_beg_day, bt_rep, bt_Y, study_day, smoother_pts, width)
}
\value{

}
\description{
Helper function for the bootstrap step.
}
