% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.edecob}
\alias{plot.edecob}
\title{Plot Event Data}
\usage{
\method{plot}{edecob}(event_data, label = "Data", ...)
}
\arguments{
\item{event_data}{The output of the \code{edecob} function. It is an object
of class \code{edecob} containing the data and the event information.}

\item{label}{The name of the y-axis in the plot.}

\item{...}{Additional parameters to be given to the function.}
}
\value{
A \code{ggplot2} object that visualizes the data.
}
\description{
Generates a ggplot2 visualizing the data and event if any detected.
}
\details{
The data points are plotted on the x-axis with the study day on the y-axis.
The data from the learning period are gray. The baseline period is outlined
by two vertical blue lines. The smoother is plotted in orange. The confidence
bound  is the blue area. If an event is detected, a red triangle will mark
the detection day.
}
