% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.edecob}
\alias{summary.edecob}
\title{Summarizing Event Detection Results}
\usage{
\method{summary}{edecob}(object, ...)
}
\arguments{
\item{object}{An object of class "\code{edecob}" for which the summary will be shown.}

\item{...}{Other arguments}
}
\value{
A short summary whether an event was detected and the parameters of the \code{edecob} function call.
}
\description{
\code{summary} method for class "\code{edecob}", gives a summary of an edecob object
}
