% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_smoother.R
\name{bt_smoother}
\alias{bt_smoother}
\title{Bootstrap the Smoother}
\usage{
bt_smoother(data, smoother, smoother_pts, resid, bt_tot_rep, ...)
}
\arguments{
\item{data}{A data frame in long format containing the data for which events
is to be detected. This means that each measurement corresponds to a row
and the columns are (in order): source (the device or person from which the
data was collected), point in time,
measurement value, lower detection bound, and upper detection bound.

The source is expected to
be a string; the time point, measurements, and detection bounds are expected to be numerical.
The detection bounds are in absolute value in the same unit as the
values and each is expected to be identical for the same source.

In case detection is wanted for a one sided change (e.g. give an event if
the confidence bounds drop below a threshold) then the upper or lower detection
bound can be chosen to be Inf or -Inf respectively.}

\item{smoother}{Which smoother is to be used. Use \code{mov_med} for the
moving median. When using the moving median, the parameter \code{width} must
be given to specify the size of the window over which the moving median is
to be taken. Defaults to the moving median.}

\item{smoother_pts}{A data frame containing the smoother with columns time_point
and value. Preferably the output of one of the smoother functions within
this package.}

\item{resid}{A vector of the same length as the number of rows of data
containing the difference between the smoother and the
measurements. Preferably the output of \code{\link{smoother_resid}}.}

\item{bt_tot_rep}{The number of iterations for the bootstrap computation. Because of
run time, it is recommended to keep this number below 500. Defaults to 100.}

\item{...}{Additional parameters to be given to the function. Possible
parameters for the model are \code{order} and \code{min_pts_in_win}. For
the moving median, a \code{width} is required.

The parameter \code{min_pts_in_win}
defines the minimal number of
measurements required to be in the time window for the median to be calculated.
Defaults to 1.

If the parameter \code{order} is given, that number will be the (maximal)
order of the autoregressive model. If no \code{order} is given, it will be
determined using the Akaike information criterion.

When the moving
median is used as the smoother, \code{width} is expected. If no \code{width} is
given, it will default to 84.}
}
\value{
A data frame containing the bootstrap repetitions of the smoother.
The column are subject identifier, time point, value, and the bootstrap
repetition the value corrsponds to.
}
\description{
First, fit an autoregressive model on the residuals of the smoother. Then bootstrap
the errors of the autoregressive model. Afterwards, reconstruct the measurements
by adding the bootstrapped error, the autoregressive model, and the smoother.
We can again calculate the smoother using these reconstructed measurements to
obtain the bootstrapped smoother (which can later be used to construct the
simultaneous confidence bounds). For details see below.
}
\details{
An autoregressive (AR) model is used for the residuals of the smoother:
\deqn{Y(t) = S(t) + \eta(t)}
\deqn{\eta(t) = \sum^{p}_{j = 1} \phiⱼ \eta(t - j) + \epsilon}
where \eqn{t} is the point in time, \eqn{Y(t)} the data point,
\eqn{S(t)} a smoother, \eqn{\eta(t)} the residual of the smoother, \eqn{p}
the order of the AR model, \eqn{\phiⱼ} the coefficients of the AR model, and
\eqn{\epsilon} the error of the AR model.

The bootstrap procedure is as follows:
\enumerate{
\item Compute the smoother \eqn{S(t)}.
\item Compute the residuals \eqn{\eta(tᵢ) = Y(tᵢ) - S(tᵢ)}.
  \item Fit an AR(p) model to \eqn{\eta(tᵢ)} to obtain the coefficients
\eqn{\phi₁, \dots, \phiₚ} and residuals \eqn{\epsilon(tᵢ) = \eta(tᵢ) -
    \sum^{p}_{j = 1} \phiⱼ \eta(tᵢ - tᵢ₋ⱼ)}.
  \item Resample \eqn{\epsilon(tᵢ)*} from \eqn{\epsilon(tₚ₊₁), \dots,
    \epsilon(tₙ)} to obtain \deqn{Y(tᵢ)* = S(tᵢ) + \eta(tᵢ)*,} where  \deqn{\eta(tᵢ)* = \sum^{p}_{j=1} \phiⱼ \eta(tᵢ₋ⱼ)*+ \epsilon(tᵢ₋ⱼ)*.}
  \item Compute \eqn{S(.)* = g(Y(t₁), \dots, Y(tₙ))} where \eqn{g} is the
function with which the smoother is calculated.
\item Repeat steps 4 and 5 \code{bt_tot_rep} times.
}
}
\references{
Bühlmann, P. (1998). Sieve Bootstrap for Smoothing in
Nonstationary Time Series. \emph{The Annals of Statistics}, 26(1), 48-83.
}
