% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_smoother.R
\name{bt_smoother}
\alias{bt_smoother}
\title{Bootstrap the Smoother}
\usage{
bt_smoother(data, smoother, smoother_pts, resid, bt_tot_rep, ...)
}
\arguments{
\item{data}{A data frame or matrix containing the data. The data points should be in the
first column, the study days in the second column, and the subject
identifier in the third column. If a matrix is used, the subject identifier
is to be specified in the row names. The study days are expected to be integer
numbers. Study day zero may be the randomization date, the first day
after baseline, or whatever is preferred. The subject identifier is expected
to be a string.}

\item{smoother}{Which smoother is to be used. Use \code{mov_med} for the
moving median. When using the moving median, the parameter \code{width} must
be given to specify the size of the window over which the moving median is
to be taken. Defaults to the moving median.}

\item{smoother_pts}{A data frame containing the smoother. Preferably the
output of one of the smoother functions included in this package.}

\item{resid}{A vector containing the residuals of the smoother to the data
points.}

\item{bt_tot_rep}{The number of times we perform the bootstrap. Defaults to 100.}

\item{...}{Additional parameters to be given to the function. When the moving
median is used, a \code{width} parameter is expected. When using the quantile
for baseline method, a \code{quantile} parameter is expected.}
}
\value{
A data frame containing the bootstrap repetitions of the smoother.
}
\description{
Fit an autoregressive model on the residuals of the smoother and bootstrap
the error of the autoregressive model. Reconstruct the smoother by adding the
bootstrapped error, the fitted autoregressive model, and the smoother to form
the bootstrapped smoother.
}
\details{
An autoregressive (AR) model is used for the residuals of the smoother:
\deqn{Y(t) = S(t) + \eta(t)}
\deqn{\eta(t) = \sum^{p}_{j = 1} \phiⱼ \eta(t - j) + \epsilon}
where \eqn{t} is the point in time, \eqn{Y(t)} the data point,
\eqn{S(t)} a smoother, \eqn{\eta(t)} the residual of the smoother, \eqn{p}
the order of the AR model, \eqn{\phiⱼ} the coefficients of the AR model, and
\eqn{\epsilon} the error of the AR model.

The bootstrap procedure is as follows:
\enumerate{
\item Compute the smoother \eqn{S(t)}.
\item Compute the residuals \eqn{\eta(tᵢ) = Y(tᵢ) - S(tᵢ)}.
  \item Fit an AR(p) model to \eqn{\eta(tᵢ)} to obtain the coefficients
\eqn{\phi₁, \dots, \phiₚ} and residuals \eqn{\epsilon(tᵢ) = \eta(tᵢ) -
    \sum^{p}_{j = 1} \phiⱼ \eta(tᵢ - tᵢ₋ⱼ)}.
  \item Resample \eqn{\epsilon(tᵢ)*} from \eqn{\epsilon(tₚ₊₁), \dots,
    \epsilon(tₙ)} to obtain \deqn{Y(tᵢ)* = S(tᵢ) + \eta(tᵢ)*,} where  \deqn{\eta(tᵢ)* = \sum^{p}_{j=1} \phiⱼ \eta(tᵢ₋ⱼ)*+ \epsilon(tᵢ₋ⱼ)*.}
  \item Compute \eqn{S(.)* = g(Y(t₁), \dots, Y(tₙ))} where \eqn{g} is the
function with which the smoother is calculated.
\item Repeat steps 4 and 5 \code{bt_tot_rep} times.
}
}
\references{
Bühlmann, P. (1998). Sieve Bootstrap for Smoothing in
Nonstationary Time Series. \emph{The Annals of Statistics}, 26(1), 48-83.
}
