% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edecob.R
\docType{package}
\name{edecob-package}
\alias{edecob-package}
\title{edecob: Event Detection Using Confidence Bounds}
\description{
Detect sustained worsening/improvement in time-dependent data.
}
\section{Main functions}{

\describe{
\item{\code{edecob}}{Detects event in the data provided}
\item{\code{edecob_plot}}{Generates a ggplot to visualize the data}
}
}

\section{Other functions}{

\describe{
\item{\code{mov_med}}{Calculates the moving median}
\item{\code{smoother_resid}}{Calculates the residuals of the moving median}
\item{\code{bt_smoother}}{Bootstraps the smoother}
\item{\code{conf_band}}{Calculates the confidence bounds}
\item{\code{detect_event}}{Detects events using the confidence bounds}
}
}

