% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edecob_plot.R
\name{edecob_plot}
\alias{edecob_plot}
\title{Plot Event Data}
\usage{
edecob_plot(
  data,
  basel,
  thresh,
  smoother_pts,
  conf_band,
  event,
  subj_id = "subj1",
  basel_start,
  basel_end,
  width,
  label = "Data"
)
}
\arguments{
\item{data}{A data frame containing the data. The values should be in the
first column, the study days in the second column, and the subject
identifier in the third column. The study days are expected to be integes
numbers. Study day zero may be the randomization date, the first day
after baseline, or whatever is preferred. The subject identifier is expected
to be a string.}

\item{basel}{The baseline for the data.}

\item{thresh}{The threshold for the data.}

\item{smoother_pts}{A data frame containing the smoother. Preferably the
output of one of the smoother functions included in this package.}

\item{conf_band}{A data frame containing the confidence bounds. Ideally the output of \code{\link{conf_band}}.}

\item{event}{A list containing the events. Preferably the output of \code{\link{detect_event}}}

\item{basel_start}{The study day of first data point to be used for the
baseline. Defaults to the first study day at which a data point is present.}

\item{basel_end}{The study day of last data point to be used for the
baseline.}

\item{label}{The name of the y-axis in the plot.}
}
\value{
A \code{ggplot2} object that visualizes the data.
}
\description{
Generates a ggplot2 visualizing the data and event if any detected.
}
\details{
The data points are plotted on the x-axis with the study day on the y-axis.
The data from the learning period are gray. The baseline period is outlined
by two vertical blue lines. The smoother is plotted in orange. The confidence
bound  is the blue area. If an event is detected, a red triangle will mark
the detection day.
}
\examples{

}
