% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edecob_plot.R
\name{edecob_plot}
\alias{edecob_plot}
\title{Plot Event Data}
\usage{
edecob_plot(event_data, label = "Data", ...)
}
\arguments{
\item{event_data}{A data frame containing the smoother. Preferably the
output of one of the smoother functions included in this package.}

\item{label}{The name of the y-axis in the plot.}

\item{...}{If data from more than one subject is included in
\code{event_data}, specify \code{subj_id}. Defaults to the first subject
found in \code{event_data$data}.}
}
\value{
A \code{ggplot2} object that visualizes the data.
}
\description{
Generates a ggplot2 visualizing the data and event if any detected.
}
\details{
The data points are plotted on the x-axis with the study day on the y-axis.
The data from the learning period are gray. The baseline period is outlined
by two vertical blue lines. The smoother is plotted in orange. The confidence
bound  is the blue area. If an event is detected, a red triangle will mark
the detection day.
}
\examples{

}
