% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Artificially generated data for examples}
\format{
A dataframe with 317 rows and 5 variables:
\describe{
\item{subject}{The subject identifier}
\item{time_point}{The point in time from which the data is collected, in days}
\item{jump_height}{The jump height, in cm}
\item{detect_lower}{The lower detection bound for the subject}
\item{detect_upper}{The upper detection bound for the subject}
}
}
\usage{
example_data
}
\description{
A dataset in the long format containing the jump height
of 3 subjects over 300 days.
}
\keyword{datasets}
