% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mov_med.R
\name{mov_med}
\alias{mov_med}
\title{Moving Median over a Time Window}
\usage{
mov_med(data, study_day, subj_id = "subj1", width = 12 * 7)
}
\arguments{
\item{data}{A vector containing the data.}

\item{study_day}{A vector containing the study days at which the data points
were collected}

\item{subj_id}{A string containing an identifier for the data.}

\item{width}{The width of the window for the moving median in days.}
}
\value{
A data frame containing the values of the moving median, the study
day to which it corresponds, the time window from which it was calculated,
and the subject id corresponding to the data.
}
\description{
Calculates the moving median over a time window around a time point for the
all time points between the first and last study day provided.
}
\details{
For the initial time points where the time difference between the first data
point and the time point for which we are calculating the median is less than
half the \code{width}, we do not have enough data points to form a window
which has the same size to both sides of the time point. In this case fewer
data points are used to calculate the median and the time window is not
symmetric around the time point for which we are calculating the median.

No median is calculated if the time difference between the last data point
and the current time point for which we are calculating the median is less
than half the \code{width}. We do not calculate the median using a smaller
time window so that the values do not change upon receiving new data.
}
