% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mov_med.R
\name{mov_med}
\alias{mov_med}
\title{Moving Median over a Time Window}
\usage{
mov_med(data, width = 12 * 7, min_pts_in_win = 1)
}
\arguments{
\item{data}{A data frame in long format containing the data for which events
is to be detected. The columns are (in order): subject identifier, study day,
measurements, baseline, and threshold. The subject identifier is expected to
be a string; the study day, measurements, baseline, and threshold are expected to be numerical.
The baseline and threshold are each expected to be identical for the same patient.}

\item{width}{The width of the window over which the moving median is taken in
number of days.}

\item{min_pts_in_win}{The minimal number of measurements required to be in
the time window in order for the median to be calculated.}
}
\value{
A data frame containing the values of the moving median, the study
day to which it corresponds, the time window from which it was calculated,
and the subject id corresponding to the data.
}
\description{
Calculates the moving median over a time window around a time point for the
all time points between the first and last study day provided.
}
\details{
Consider a sample \eqn{X₁,\dots, Xₙ} of size \eqn{n} and the
reordering OcRMtgPou0fgW7qwAKGTL5opXj7i1JW7-3\eqn{X₍₁₎ \le X₍₂₎ \le \dots \le X₍ₙ₎}, commonly
called the order statistic. Then for \eqn{n} even the median usually
defined as \deqn{median(X₁,\dots, Xₙ) = X₍ₖ₎, where k = n/2.} In the
case where \eqn{n} is odd the median is
defined as \deqn{median(X₁,\dots, Xₙ) = 1/2(X₍ₖ₎ + X₍ₖ₊₁₎), where k = n/2.} Let the
study days at which the measurements \eqn{X₁, \dots, Xₙ} were taken
be \eqn{t₁, \dots, tₙ}.
Let \eqn{T} a fixed positive amount of time. Then the moving median at time
point \eqn{t} with window size \eqn{T} is defined as
\deqn{S(t) = median({Xⱼ | t - T/2 \le tⱼ \le t + T/2}).}

For the initial time points where the time difference between the first data
point and the time point for which we are calculating the median is less than
half the \code{width}, we do not have enough data points to form a window
which has the same size to both sides of the time point. In this case fewer
data points are used to calculate the median and the time window is not
symmetric around the time point for which we are calculating the median.

No median is calculated if the time difference between the last data point
and the current time point for which we are calculating the median is less
than half the \code{width}. We do not calculate the median using a smaller
time window so that the values do not change upon receiving new data.
}
