% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.edecob}
\alias{plot.edecob}
\title{Plot Event Data}
\usage{
\method{plot}{edecob}(x, ...)
}
\arguments{
\item{x}{The output of the \code{edecob} function for one subject. It is an object
of class \code{edecob} containing the data and the event information.}

\item{...}{Other arguments like \code{title}, \code{xlab}, or \code{ylab}. If the
plot should be returned, write \code{output = TRUE}.}
}
\value{
None. If \code{output = TRUE} was written in function call, a \code{ggplot}
object that visualizes the data will be returned. The returned plot will not
contain the the text at the bottom.
}
\description{
Generates a ggplot2 visualizing the data and event if any detected.
}
\details{
The data points are plotted on the x-axis with the time point on the y-axis.
The data from the learning period are gray. The baseline period is outlined
by two vertical blue lines. The smoother is plotted in orange. The confidence
bound  is the blue area. If an event is detected, a red triangle will mark
the detection day.

If the parameter \code{output = TRUE} is given to the function, the function
will output the ggplot. The plot can then be manipulated using the usual functions
from the ggplot2 package.
}
