% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoother_resid.R
\name{smoother_resid}
\alias{smoother_resid}
\title{Residuals of the Smoother}
\usage{
smoother_resid(data, study_day, smoother_pts)
}
\arguments{
\item{data}{A data frame containing the data. The values should be in the
first column, the study days in the second column, and the subject
identifier in the third column. The study days are expected to be integes
numbers. Study day zero may be the randomization date, the first day
after baseline, or whatever is preferred. The subject identifier is expected
to be a string.}

\item{smoother_pts}{A data frame containing the smoother. Preferably the
output of one of the smoother functions included in this package.}
}
\value{
A vector of the same length as \code{data} containing the residuals.
}
\description{
Calculate the residuals of the smoother to the data points.
}
