% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoother_resid.R
\name{smoother_resid}
\alias{smoother_resid}
\title{Residuals of the Smoother}
\usage{
smoother_resid(data, study_day, smoother_pts)
}
\arguments{
\item{data}{A vector containing the data.}

\item{study_day}{A vector containing the study days at which the data points
were collected}
}
\value{
A vector of the same length as \code{data} containing the residuals.
}
\description{
Calculate the residuals of the smoother to the data points.
}
