% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoother_resid.R
\name{smoother_resid}
\alias{smoother_resid}
\title{Residuals of the Smoother}
\usage{
smoother_resid(data, smoother_pts)
}
\arguments{
\item{data}{A data frame in long format containing the data for which events
is to be detected. This means that each measurement corresponds to a row
and the columns are (in order): subject identifier, point in time,
measurement value, baseline, and threshold.

The subject identifier is expected to
be a string; the time point, measurements, baseline, and threshold are expected to be numerical.
The baseline and threshold are in absolute values in the same unit as the
measurements and each is expected to be identical for the same patient. The
threshold can be above or below the baseline. For details see below.}

\item{smoother_pts}{A data frame containing the smoother. Preferably the
output of one of the smoother functions included in this package.}
}
\value{
A vector of the same length as \code{data} containing the residuals.
}
\description{
Calculate the residuals of the smoother to the data points.
}
